requirements.txt
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
backcall==0.2.0
beautifulsoup4==4.11.1
decorator==5.1.1
Deprecated==1.2.13
Django==4.1
django-appconf==1.0.5
django-debug-toolbar==3.6.0
django-filter==22.1
djangorestframework==3.14.0
executing==0.10.0
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.4.0
traitlets==5.3.0
wcwidth==0.2.5
wrapt==1.14.1
